Rename migration files, delete channel connection table and change channel id type to smallint.

+++ backend/src/main/resources/db/migration/V3__Create_program_table.sql
@@ -5,6 +5,6 @@ create table program (
     start timestamp not null,
     end timestamp not null,
     content_id bigint not null,
-    channel_id bigint not null,
+    channel_id smallint not null,
     primary key (id)
 );
